marked kernel in mm_table

+++ include/mm/physical.h
@@ -8,8 +8,8 @@
 #define MM_PHYSICAL_MAX 64
 #define MM_TABLE_MAX (((uint64_t) MM_PHYSICAL_MAX * 1024 * 1024 * 1024) / (8 * sizeof(uint64_t) * PG_SIZE))
 
-extern void physical_free_range(uint64_t begin, size_t length);
-extern void physical_take_range(uint64_t begin, size_t length);
+extern void physical_free_range(uint64_t begin, uint64_t length);
+extern void physical_take_range(uint64_t begin, uint64_t length);
 extern void physical_set_limit(uint64_t limit);
 
 #endif
